% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disequilibrium.R
\docType{data}
\name{fjdata}
\alias{fjdata}
\title{Data from Fair and Jaffee (1972).}
\format{An object of class \code{data.frame} with 127 rows and 8 columns.}
\usage{
fjdata
}
\description{
A data set of 126 monthly observations of the housing market from June 1959 to November 1969. The variables are

\eqn{T}: Time. A continuous values time variable
\eqn{\sum_{i=1}^{t-1}HS_i}{\sum[i=1]^{t-1}HS[i]}: Stock of houses. A sum of housing starts over all previous periods. Assuming initial stock is 0.
\eqn{RM_{t-2}}{RM[t-2]}: Mortgage rate lagged by two months.
\eqn{DF6_{t-1}}{DF6[t-1]}: A six month moving average of $DF_t$. $DF_t$ is the flow of private deposits into savings and loan associations (SLA) and mutual savings banks during period $t$.
\eqn{DHF3_{t-2}}{DHF3[t-2]}: The flow of borrowings by the SLAs from the federal home-loan bank during month t.
\eqn{RM_{t-1}}{RM[t-1]}: Mortgage rate lagged by one month.
}
\references{
Fair, Ray C., and Dwight M. Jaffee. "Methods of estimation for markets in disequilibrium." Econometrica: Journal of the Econometric Society (1972): 497-514.
}
\keyword{datasets}
