% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disequilibrium.R
\name{GetDeltaMethodParameters}
\alias{GetDeltaMethodParameters}
\title{GetDeltaMethodParameters}
\usage{
GetDeltaMethodParameters(mu, covmat = NULL, MaskRho = FALSE)
}
\arguments{
\item{mu}{A numeric vector output from DE()$par}

\item{covmat}{A numeric matrix output from DE()$vcov}

\item{MaskRho}{The output from DE()$MaskRho}
}
\value{
The parameters of the normal distribution after transforming the coavariance variables to a positive definite space
}
\description{
Transforms the mean and variance covariance matrix of the estimators in an unrestricted space to a positive definite space for the delta method.
}
\examples{
GetDeltaMethodParameters
}
