% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disequilibrium.R
\name{TransformSigma_PDtoR2}
\alias{TransformSigma_PDtoR2}
\title{TransformSigma_PDtoR2}
\usage{
TransformSigma_PDtoR2(vec)
}
\arguments{
\item{vec}{A length 2 vector of the variance of equation 1 followed by the variance of equation 2.}
}
\value{
A length 2 vector spanning unrestricted R2.
}
\description{
TransformSigma_PDtoR2
}
\examples{
PD_vec <- c(1, 1)
TransformSigma_PDtoR2(PD_vec)

}
