% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disequilibrium.R
\name{TransformSigma_PDtoR3}
\alias{TransformSigma_PDtoR3}
\title{TransformSigma_PDtoR3}
\usage{
TransformSigma_PDtoR3(vec)
}
\arguments{
\item{vec}{A length 3 vector of the variance of equation 1, followed by the
covariance of equations 1 and 2, followed by the variance of equation 2.}
}
\value{
A length 3 vector spanning unrestricted R3.
}
\description{
TransformSigma_PDtoR3
}
\examples{
PD_vec <- c(1, 0, 1)
TransformSigma_PDtoR3(PD_vec)

}
