% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disequilibrium.R
\name{TransformSigma_R2toPD}
\alias{TransformSigma_R2toPD}
\title{TransformSigma_R2toPD}
\usage{
TransformSigma_R2toPD(vec)
}
\arguments{
\item{vec}{A length 2 vector output from TransformSigma_PDtoR2.}
}
\value{
A length 2 vector of variances spanning a positive definite space.
}
\description{
TransformSigma_R2toPD
}
\examples{
PD_vec <- c(1, 1)
R2_vec <- TransformSigma_PDtoR2(PD_vec)
TransformSigma_R2toPD(R2_vec)

}
