% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disequilibrium.R
\name{TransformSigma_R3toPD}
\alias{TransformSigma_R3toPD}
\title{TransformSigma_R3toPD}
\usage{
TransformSigma_R3toPD(vec)
}
\arguments{
\item{vec}{A length 3 vector output from TransformSigma_PDtoR3.}
}
\value{
A length 3 vector spanning a positive definite space.
}
\description{
TransformSigma_R3toPD
}
\examples{
PD_vec <- c(1, 0, 1)
R3_vec <- TransformSigma_PDtoR3(PD_vec)
TransformSigma_R3toPD(R3_vec)

}
